Relocate support for reading JSON test resources.
